backend/requirements.txt
fastapi
uvicorn
python-multipart
langchain
langchain-community
faiss-cpu
sentence-transformers
edge-tts
SpeechRecognition
pydub
python-dotenv
openai
